Hotkey handling, endge cases
