+++ package.json
@@ -81,6 +81,7 @@
     "react-gtm-module": "^2.0.11",
     "react-helmet": "^6.1.0",
     "react-hook-form": "^7.48.2",
+    "react-hotkeys-hook": "^5.1.0",
     "react-infinite-scroll-component": "^6.1.0",
     "react-inlinesvg": "^3.0.2",
     "react-intl": "^5.21.2",
